Implement Non-Destructive Plugin Manager
Add plugin manager that allows any number of plugins to inject code
anywhere that a hook is established. Includes mechanism so that if any
plugin returns true then the default behavior can be overriden depending
on how the hook is implemented.

The only hooks at the moment are
-login (Player connects, no override)
-logout (Player disconnects, no override)
-parse (New line of server output)
-chat (Player sends a chat message or command)

An example plugin can be found in the ./plugins/ directory